[SKB-2176] Add Image in SG goes to 2nd screen (#19)
* [SKB-2176] Add Image in SG goes to 2nd screen

* Monkey patch QToolButton and QMenu to resolve Qt 5.15.0 bug (QTBUG-84462)
* Override QMenu showEvent to reposition the menu correctly before showing it.

* Only reposition menus that are showing from QToolButtons.

* Add patch for QToolButton addAction.

QToolButton menus can be created at the time of popup, if QActions are added to the QToolButton, instead of using the setMenu method. Override the addAction method in QToolButtonPatch to make sure we create the menu on the python side, and not in the C++ source, to ensure we're using the QMenuPatch object.

Add extra handling for finding the menu orientation to help reposition it.